requirements.txt
Flask
python-dotenv
pyttsx3
google-generativeai==0.8.5
google-search-results
gunicorn
